[auth] introduce passthrough option (#68)
new config.auth.passthrough: bool setting. Default False => current
behaviour. But if True, changes the `get_active_user` method to
allow for returning None. The execute methods should be happy with
None user, and admin methods were changed to allow for None in
case the config.auth.passthrough is True

the fiab.sh bootstrap script sets the passthrough=True, as it is
intended for the single-user standalone local regime