Introduce pylock.toml
We introduce a pylock.toml file, to capture the environment exactly,
so that we can have reproducible installs.

This pylock.toml gets created during a release, where we test a clean
install with the just-built package and this lock. However, this
requires a rework of the install process, so that we actually commit
the pylock.toml in the commit that is tagged.

We then make some minor changes to the fiab.sh, in particular,
respecting this pylock.toml, downloading it, noting the version/timestamp,
downloading base config file, etc.

We remove the FIAB_DEV option from fiab.sh -- for development, one should
simply use `uv sync`, `uv run, etc.

Updates of the pylock.toml are still driven by the pyproject.toml/uv.lock.
If we want to update some package, say, multiolib -- we would either do it
in the pyprojec.toml, or by `uv lock --upgrade`. The former applies to
hard constraints (like known incompatibilities or new api enabled),
the latter for regular "lets just use a newer version".